Implemented clients for the services to use
I realized that I was badly breaking abstractions by having the service return a
type that was used directly by the UI code. This resulted in a coupling that
made it difficult to change the shape of the data model without impacting the
UI. In particular, I needed to start recording the time that the customer was
added to the line so I could properly determine if their patience had run out. I
needed to store the time they entered the line. Since I was using just one
object, that change effectively broke the UI code. The UI code thus far doesn't
actually care about the time the customer entered the line, and shouldn't have
been affected by this change.

The solution here was to separate the service functionality into a client.
Starting with the cashier, customer and line services, each of these services
use a client to perform the intended operation. Additionally, we do a
transformation from the remote type to a local type. This breaks the coupling
between our middlewares and the remote types.

We will need to ask the line service when the customer entered the line, then we
pass that time to the clock service to get it normalized, so we know how many
seconds have elapsed since then.