Phase 1.1 - Remove service layer
Intially, I thought of the service layer as a local abstraction of the client
that we're calling to get/update data for customers, cashiers, lines, etc... I'm
actually not seeing the value of this abstraction though.

The client should provide sufficient abstraction over the minutiae of the
backend. The main change I need to make is to have the client return local types
instead of remote ones. By returning local types, it means our business logic
dictates the types that are useful for our application, as opposed to types that
are convenient for the backend to return. This also ensures that when we change
the client implementation, we can maintain the same contract with the rest of
our code.

As such, in this task, I'm going to surgically remove the services layer,
leaving only a client layer. The client will now simulate a backend API layer.
Later we'll likely move this to an actual backend. That will be the true test of
utility of the abstraction, and the importance of having our business logic
determine the contract of the client.

One more change -- I added numbers to each of the tasks so they can be tracked
more easily and I can create branches and perhaps tags when I've completed the
task.