Removed services abstraction layer
In this commit I'm removing the services abstraction layer. Previously I had a
service calling a client for each entity (e.g. customer service called a
customer client). I found that the service abstraction didn't add much value. It
was just an arbitrary layer. I decided instead to modify the clients to operate
on local types instead of remote ones.

Therefore, in this commit, I'm removing the services and changing the client to
hide the remote types and only return local types to the caller. This ensures
that we can easily change the client without modifying the rest of our code.
Later we may want to use an SQLite database, or JSON database, or make API
calls.

In the process of ensuring that the client's public interface uses local types I
had to add some internal getRemote and updateRemote methods. This was a tad
painful since I'm using an internal array or mapping to store the data for now.

I had to add some missing RemoteXXX types to facilitate the clean separation
between local and remote types. Remote represents the internal serialization
format that the client uses.

Finally, I had to add typescript as a dependency to run React. I'm not sure why
that was necessary. I got the error trying to run the project after doing "yarn
install" on a different machine. I'm not sure what changed.