Phase 1.4 - Add tests for clients
I haven't written any tests for any of the code as yet, so I'm going to start
writing tests for the client. These tests will ensure that we're properly
persisting and returning data for the UI. Also, it will provide some insight on
how the tests will change once we start making changes to the clients.