Added Architecture Decision Records (ADRs)
I decided to catalogue architectural decisions using ADRs. I added a template to
the repo and used it for two decisions.

Advancing the clock in tests
The Engine is heavily reliant on clock manipulation in tests. I suspect, other
areas of the system will require the clock as well. I needed to determine
whether each test should use Jest's tools to manipulate the clock, or we should
provide an API abstraction.

Customer Line management API
I needed to determine whether the Line API facilitated the creation/management
of customers in the line. Or, do we have a separate API for managing the
customers in a given line?