Simplified fetching of customers in line
In building out the functionality to access the customers
in a line, I ran into some complications. The `CustomerInLine`
object only provided a reference to the customer's ID.
This made working with the `Line` and `CustomerInLine` types
more complicated.

For the UI, I wanted to list some properties of the customer
for the line they were listed in. Not having the customer
nested on the `CustomerInLine` object meant, I'd need to
find a way to get that customer per `CustomerInLine` object.

Also, `useSelector` doesn't allow us to do anything fancy,
so I can't select the customer as I'm looping through. Also,
selecting data while rendering isn't something we want to do.
The render process happens multiple times and we don't want
to incur that overhead.

So.. I nested it under the `CustomerInLine` object. One
idea worth exploring is whether I could have kept the clients
simpler and instead introduced a Data layer of some kind
whose responsibility is to bring these pieces of data
together.

Effectively, our clients are playing multiple roles currently.
It's fetching data from our "backend", but it's also massaging
data into the format that's best suited for our application.