Fixed MAJOR bug with testing
Every time I ran a test that had something
to do with the clock starting, the Jest
suite would fail in a way that wasn't
related to any of the tests.

It would complain that customer
I was trying to access didn't exist.

I recalled quite a few things about
how some aspects of the app works.
- The test runs pending timers at
  the end. I noticed that a test
  that should have only advanced the
  clock by one tick advanced it by
  2 ticks instead. This happens
  because our global Jest config
  runs pending timers. This is
  the recommended approach to avoid
  weird idiosyncrasies.
  This was causing the 2nd tick to
  get processed.
- The second issue was that the Line
  Client was defining its own
  CustomerClient. Since we're storing
  local state in the client, that's
  actually a big deal. Since the
  LineClient declared its own
  CustomerClient, that meant it
  had no customers in it. Because
  it had no customers, we get the
  error that we did in our tests.