Debugged Engine.spec.ts broken test due to promises.
The Engine test was broken in a really obtuse way.
The addCustomersToLines method was awaiting the
creation of assignment of the customer to the line,
then it removed the customer from the queue later
in the same method.

This didn't work. The removal of the customer from
the queue was executed AFTER the rest of the test
ran.

This happens because the remaining work in the method
was queued to be executed after the rest of the test.
When we await something, the rest of the function
is wrapped in a "then" clause basically, which gets
added to the microtask queue.

The microtask queue is only visited once the call stack
is empty. Once the remaining method body (addCustomersToLines)
was added to the microtask queue, Javascript
went back to the test and executed the subsequent
lines.

One open question is why the microtask was eventually
executed. Was it done at the end of the test?